Парсеры поделены на модули
