+++ LiteMD/LiteMD.vcxproj.filters
@@ -136,17 +136,20 @@
     <ClCompile Include="string_processing\crlfParser.cpp">
       <Filter>String processing\Парсеры</Filter>
     </ClCompile>
-    <ClCompile Include="string_processing\urlBasicParser.cpp">
+    <ClCompile Include="string_processing\urlBasicSimplifiedParser.cpp">
       <Filter>String processing\Парсеры</Filter>
     </ClCompile>
     <ClCompile Include="string_processing\regex.cpp">
       <Filter>String processing</Filter>
     </ClCompile>
-    <ClCompile Include="string_processing\urlAlternateParser.cpp">
+    <ClCompile Include="string_processing\urlAdvancedParser.cpp">
       <Filter>String processing\Парсеры</Filter>
     </ClCompile>
     <ClCompile Include="string_processing\symbolCleaner.cpp">
       <Filter>String processing\Парсеры</Filter>
+    </ClCompile>
+    <ClCompile Include="string_processing\urlBasicParser.cpp">
+      <Filter>String processing</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -284,16 +287,19 @@
     <ClInclude Include="string_processing\crlfParser.h">
       <Filter>String processing\Парсеры</Filter>
     </ClInclude>
-    <ClInclude Include="string_processing\urlBasicParser.h">
+    <ClInclude Include="string_processing\urlBasicSimplifiedParser.h">
       <Filter>String processing\Парсеры</Filter>
     </ClInclude>
     <ClInclude Include="string_processing\regex.h">
       <Filter>String processing</Filter>
     </ClInclude>
-    <ClInclude Include="string_processing\urlAlternateParser.h">
+    <ClInclude Include="string_processing\urlAdvancedParser.h">
       <Filter>String processing\Парсеры</Filter>
     </ClInclude>
     <ClInclude Include="string_processing\symbolCleaner.h">
+      <Filter>String processing\Парсеры</Filter>
+    </ClInclude>
+    <ClInclude Include="string_processing\urlBasicParser.h">
       <Filter>String processing\Парсеры</Filter>
     </ClInclude>
   </ItemGroup>
