+++ LiteMD/string_processing/urlBasicParser.h
@@ -1,5 +1,1 @@
 #pragma once
-#include "regex.h"
-#include <string>
-
-std::wstring basicUrlParser(std::wstring& rawInput);
