+++ LiteMD/string_processing/urlBasicSimplifiedParser.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "regex.h"
+#include <string>
+
+std::wstring basicUrlParser(std::wstring& rawInput);
