+++ LiteMD/testPattern.md
@@ -1,1 +1,1 @@
-t2>> j>v5<<5<>* (#@%*(% >^<www.google.com>6 ,>^>^< <<[google](www.google.com)@<#>$%^>@<>^7 >^&<&<78 > 
+t2>> j>v5<<5<>* (#@%*(% >^<www.google.com>6 ,>^>^<кукареку<https://www.yandex.ru><-жопытри <<[google](www.google.com)@<#>$%^>@<>^7 >^&<&<78 >
