+++ LiteMD/ver.ini
@@ -1,2 +1,2 @@
 [VersionInformation]
-Current Build Number=1326
+Current Build Number=1337
